refactor: extract VmExpressionEvaluator and VmStatementExecutor from VirtualMachine
Extract expression evaluation (33 methods) into VmExpressionEvaluator and
statement execution (17 methods) into VmStatementExecutor as package-private
final delegate classes. VirtualMachine shrinks from 1193 to 475 lines.

Key changes:
- VmExpressionEvaluator: 459 lines, owns evaluate() switch dispatch and all
  evaluateX methods, evaluateBoolean/evaluateInt are package-private for
  cross-delegate access from executor
- VmStatementExecutor: 376 lines, owns execute() switch dispatch and all
  executeX methods
- virtualMachineListeners upgraded from synchronized LinkedList to
  CopyOnWriteArrayList (removes 5 synchronized blocks)
- isStopped, virtualMachineListeners, mapAbstractClsToAdapterCls, checkNotNull
  promoted to package-private for delegate access
- evaluate(), execute(), evaluateArguments() remain on VirtualMachine as thin
  wrappers preserving the public/protected API
- ReleaseVirtualMachine unchanged
- No files outside core/ast modified

Closes #551
